+++ src/Modules/Changelog.lua
@@ -3,6 +3,22 @@ local _, addon = ...
 local Changelog = {}
 Changelog.list = {
   {
+    version = "12.0.5.1",
+    date = "2026-04-27",
+    notify = true,
+    categories = {
+      ["New"] = {
+        "Added option to configure desired font outline for the displayed text.",
+        "Added option to configure background color, transparency and padding for the displayed text.",
+        "Added option to choose cursor anchor direction and distance for the text positioning.",
+      },
+      ["Improvements"] = {
+        "Improved setting logs for sliders by debouncing output to avoid chat spam during dragging.",
+        "Improved setting tooltips to include the default value for the setting.",
+      },
+    }
+  },
+  {
     version = "12.0.5.0",
     date = "2026-04-24",
     notify = false,